# openapi.json
{"openapi": "3.1.0", "info": {"title": "FinControl API", "version": "1.0.0"}, "paths": {"/users/phone/{phone}": {"get": {"tags": ["users"], "summary": "Get User By Phone", "operationId": "get_user_by_phone_users_phone__phone__get", "parameters": [{"name": "phone", "in": "path", "required": true, "schema": {"type": "string", "title": "Phone"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/register": {"post": {"tags": ["users"], "summary": "Register User", "operationId": "register_user_users_register_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}}}}, "/users/": {"get": {"tags": ["users"], "summary": "Read Users", "operationId": "read_users_users__get", "parameters": [{"name": "skip", "in": "query", "required": false, "schema": {"type": "integer", "default": 0, "title": "Skip"}}, {"name": "limit", "in": "query", "required": false, "schema": {"type": "integer", "default": 100, "title": "Limit"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/UserOut"}, "title": "Response Read Users Users  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/{user_id}": {"get": {"tags": ["users"], "summary": "Read User", "operationId": "read_user_users__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["users"], "summary": "Update User", "operationId": "update_user_users__user_id__put", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["users"], "summary": "Delete User", "operationId": "delete_user_users__user_id__delete", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/{user_id}/profile": {"post": {"tags": ["users"], "summary": "Create User Profile", "operationId": "create_user_profile_users__user_id__profile_post", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProfileCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/check/{phone}": {"get": {"tags": ["users"], "summary": "Check User By Phone", "operationId": "check_user_by_phone_users_check__phone__get", "parameters": [{"name": "phone", "in": "path", "required": true, "schema": {"type": "string", "title": "Phone"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/profiles/all": {"get": {"tags": ["users"], "summary": "List All Profiles", "operationId": "list_all_profiles_users_profiles_all_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/whatsapp/webhook": {"post": {"tags": ["whatsapp"], "summary": "Webhook", "operationId": "webhook_whatsapp_webhook_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/WebhookRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/WebhookResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/whatsapp/send": {"post": {"tags": ["whatsapp"], "summary": "Send Message", "operationId": "send_message_whatsapp_send_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/Message"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/whatsapp/qr": {"get": {"tags": ["whatsapp"], "summary": "Get Qr", "operationId": "get_qr_whatsapp_qr_get", "responses": {"200": {"description": "Successful Response", "content": {"text/html": {"schema": {"type": "string"}}}}}}}, "/whatsapp/auth": {"get": {"tags": ["whatsapp"], "summary": "Get Whatsapp Auth", "operationId": "get_whatsapp_auth_whatsapp_auth_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}, "post": {"tags": ["whatsapp"], "summary": "Save Whatsapp Auth", "operationId": "save_whatsapp_auth_whatsapp_auth_post", "requestBody": {"content": {"application/json": {"schema": {"type": "object", "title": "Data"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/whatsapp/financial-context/{user_id}": {"get": {"tags": ["whatsapp"], "summary": "Get Financial Context", "operationId": "get_financial_context_whatsapp_financial_context__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/transactions/user/{user_id}": {"get": {"tags": ["transactions"], "summary": "Get User Transactions", "operationId": "get_user_transactions_transactions_user__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/TransactionOut"}, "title": "Response Get User Transactions Transactions User  User Id  Get"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/transactions/expense": {"post": {"tags": ["transactions"], "summary": "Create Expense", "operationId": "create_expense_transactions_expense_post", "parameters": [{"name": "user-id", "in": "header", "required": true, "schema": {"type": "string", "title": "User-Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/transactions/income": {"post": {"tags": ["transactions"], "summary": "Create Income", "operationId": "create_income_transactions_income_post", "parameters": [{"name": "user-id", "in": "header", "required": true, "schema": {"type": "string", "title": "User-Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/transactions/{transaction_id}": {"delete": {"tags": ["transactions"], "summary": "Delete Transaction", "operationId": "delete_transaction_transactions__transaction_id__delete", "parameters": [{"name": "transaction_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Transaction Id"}}, {"name": "user-id", "in": "header", "required": true, "schema": {"type": "string", "title": "User-Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["transactions"], "summary": "Update Transaction", "operationId": "update_transaction_transactions__transaction_id__put", "parameters": [{"name": "transaction_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Transaction Id"}}, {"name": "user-id", "in": "header", "required": true, "schema": {"type": "string", "title": "User-Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionCreate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TransactionOut"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/categories/": {"post": {"tags": ["categories"], "summary": "Create Category", "operationId": "create_category_categories__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CategoryCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/categories/user/{user_id}": {"get": {"tags": ["categories"], "summary": "Get User Categories", "operationId": "get_user_categories_categories_user__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}, {"name": "type", "in": "query", "required": false, "schema": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Type"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/token": {"post": {"tags": ["auth", "auth"], "summary": "Login", "operationId": "login_auth_token_post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_auth_token_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Token"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/reports/summary/{user_id}": {"get": {"tags": ["reports"], "summary": "Get Financial Summary", "operationId": "get_financial_summary_reports_summary__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}, {"name": "start_date", "in": "query", "required": false, "schema": {"type": "string", "title": "Start Date"}}, {"name": "end_date", "in": "query", "required": false, "schema": {"type": "string", "title": "End Date"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/reports/monthly/{user_id}": {"get": {"tags": ["reports"], "summary": "Get Monthly Report", "operationId": "get_monthly_report_reports_monthly__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}, {"name": "year", "in": "query", "required": false, "schema": {"type": "integer", "title": "Year"}}, {"name": "month", "in": "query", "required": false, "schema": {"type": "integer", "title": "Month"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/": {"get": {"summary": "Root", "operationId": "root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}, "head": {"summary": "Head Root", "operationId": "head_root__head", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"Body_login_auth_token_post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_auth_token_post"}, "CategoryCreate": {"properties": {"name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}, "icon": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Icon"}, "color": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Color"}, "user_id": {"type": "string", "title": "User Id"}}, "type": "object", "required": ["name", "type", "user_id"], "title": "CategoryCreate"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Message": {"properties": {"to": {"type": "string", "title": "To"}, "text": {"type": "string", "title": "Text"}}, "type": "object", "required": ["to", "text"], "title": "Message"}, "MessageType": {"type": "string", "enum": ["INCOME", "EXPENSE", "BALANCE", "STATEMENT", "REPORT", "CATEGORIES", "CATEGORY_REPORT", "ERROR", "HELP_MESSAGE", "HELP_DETAILS", "FINANCIAL_ADVICE"], "title": "MessageType"}, "PeriodType": {"type": "string", "enum": ["DAILY", "WEEKLY", "MONTHLY", "YEARLY"], "title": "PeriodType"}, "ProfileCreate": {"properties": {"phone": {"type": "string", "title": "Phone", "description": "Phone number in international format", "example": "5511953238980"}}, "type": "object", "required": ["phone"], "title": "ProfileCreate"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type", "default": "bearer"}}, "type": "object", "required": ["access_token"], "title": "Token"}, "TransactionCreate": {"properties": {"amount": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Amount", "description": "Valor da transa\u00e7\u00e3o", "examples": ["50.00", "1000.00"]}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description", "description": "Descri\u00e7\u00e3o da transa\u00e7\u00e3o"}, "category": {"type": "string", "title": "Category", "description": "Categoria da transa\u00e7\u00e3o"}, "date": {"type": "string", "format": "date-time", "title": "Date", "description": "Data da transa\u00e7\u00e3o"}}, "type": "object", "required": ["amount", "category"], "title": "TransactionCreate"}, "TransactionOut": {"properties": {"amount": {"type": "string", "title": "Amount", "description": "Valor da transa\u00e7\u00e3o", "examples": ["50.00", "1000.00"]}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description", "description": "Descri\u00e7\u00e3o da transa\u00e7\u00e3o"}, "category": {"type": "string", "title": "Category", "description": "Categoria da transa\u00e7\u00e3o"}, "date": {"type": "string", "format": "date-time", "title": "Date", "description": "Data da transa\u00e7\u00e3o"}, "id": {"type": "string", "title": "Id"}, "user_id": {"type": "string", "title": "User Id"}, "type": {"type": "string", "title": "Type"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}, "bankAccountId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Bankaccountid"}, "categoryId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Categoryid"}, "accountId": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Accountid"}}, "type": "object", "required": ["amount", "category", "id", "user_id", "type", "created_at", "updated_at"], "title": "TransactionOut"}, "UserOut": {"properties": {"id": {"type": "string", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "phone": {"type": "string", "title": "Phone"}, "role": {"$ref": "#/components/schemas/UserRole"}, "is_active": {"type": "boolean", "title": "Is Active"}, "created_at": {"type": "string", "format": "date-time", "title": "Created At"}, "updated_at": {"type": "string", "format": "date-time", "title": "Updated At"}}, "type": "object", "required": ["id", "email", "name", "phone", "role", "is_active", "created_at", "updated_at"], "title": "UserOut"}, "UserResponse": {"properties": {"id": {"type": "string", "title": "Id"}, "email": {"type": "string", "title": "Email"}, "name": {"type": "string", "title": "Name"}, "phone": {"type": "string", "title": "Phone"}}, "type": "object", "required": ["id", "email", "name", "phone"], "title": "UserResponse"}, "UserRole": {"type": "string", "enum": ["ADMIN", "USER"], "title": "UserRole"}, "UserUpdate": {"properties": {"name": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Name"}, "email": {"anyOf": [{"type": "string", "format": "email"}, {"type": "null"}], "title": "Email"}, "password": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Password"}, "phone": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Phone"}, "is_active": {"anyOf": [{"type": "boolean"}, {"type": "null"}], "title": "Is Active"}}, "type": "object", "title": "UserUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}, "WebhookRequest": {"properties": {"type": {"$ref": "#/components/schemas/MessageType"}, "phone": {"type": "string", "title": "Phone"}, "user_id": {"type": "string", "title": "User Id"}, "amount": {"anyOf": [{"type": "number"}, {"type": "string"}, {"type": "null"}], "title": "Amount"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "category": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Category"}, "period": {"anyOf": [{"$ref": "#/components/schemas/PeriodType"}, {"type": "null"}], "default": "MONTHLY"}, "transaction_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Transaction Id"}, "date": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Date"}, "financialContext": {"anyOf": [{"type": "object"}, {"type": "null"}], "title": "Financialcontext"}}, "type": "object", "required": ["type", "phone", "user_id"], "title": "WebhookRequest"}, "WebhookResponse": {"properties": {"message": {"type": "string", "title": "Message"}, "success": {"type": "boolean", "title": "Success", "default": true}}, "type": "object", "required": ["message"], "title": "WebhookResponse"}}}}